Use (long long) so we match the type of %ll -- int64 and uint64 may well be (long int) on some 64 bit platforms, which causes a warning.

+++ core_e/zoolib/HTTP.cpp
@@ -167,7 +167,7 @@ void Response::Set(const string& iName, int iValue)
 
 void Response::Set(const string& iName, uint64 iValue)
 	{
-	fHeaders.push_back(pair<string, string>(iName, sStringf("%lld", iValue)));
+	fHeaders.push_back(pair<string, string>(iName, sStringf("%lld", (long long)iValue)));
 	}
 
 void Response::Send(const ChanW_Bin& iChanW) const
